outra edição

+++ commands/socialcommands/say.js
@@ -1,30 +1,47 @@
-const Neable = require('c:/Brie/neable_module/NeableCommands');
+const Neable = require('../../neable_module/NeableCommands');
 module.exports.run = async (Brie, message, args) => {
 
+    if (args[0] == 'embed') {
+        data = require('../../embeds.json');
+        
+        data = JSON.stringify(data);
+        
+        data = JSON.parse(data);
+        
+        data.savedEmbeds.forEach(embed => {
+            
+            if (embed.EmbedID == args[1]) {
+                message.delete();
+                return message.channel.send({ embed: embed });
+                
+            }
+        })
+    } else {
 
-    TestarURL = new RegExp(/^(?:http(s)?:\/\/)?[\w.-]+(?:\.[\w\.-]+)+[\w\-\._~:/?#[\]@!\$&'\(\)\*\+,;=.]+$/gm)
+        TestarURL = new RegExp(/^(?:http(s)?:\/\/)?[\w.-]+(?:\.[\w\.-]+)+[\w\-\._~:/?#[\]@!\$&'\(\)\*\+,;=.]+$/gm)
 
-    messageContent = args.slice(0).join(" ");
+        messageContent = args.slice(0).join(" ");
 
-    if (message.attachments.first()) return message.reply(`ERROR 401: Eu não posso repetir arquivos!`);
+        if (message.attachments.first()) return message.reply(`ERROR 401: Eu não posso repetir arquivos!`);
 
-    if (message.content.includes('@everyone') || message.content.includes('@here')) return message.reply(`ERROR 505: Eu não posso todos do servidor!`);
+        if (message.content.includes('@everyone') || message.content.includes('@here')) return message.reply(`ERROR 505: Eu não posso todos do servidor!`);
 
-    if (message.mentions.roles.first()) return message.reply('Heey, eu não posso mencionar cargos!');
+        if (message.mentions.roles.first()) return message.reply('Heey, eu não posso mencionar cargos!');
 
-    if (!messageContent) return message.reply(`ERROR 404: Não encontrei o que eu tenho que dizer...`)
+        if (!messageContent) return message.reply(`ERROR 404: Não encontrei o que eu tenho que dizer...`)
 
-    TestarURL.test(messageContent) ? err() : say(messageContent)
+        TestarURL.test(messageContent) ? err() : say(messageContent)
 
-    function say(messageContent) {
-        message.delete();
-        message.channel.send(messageContent);
+        function say(messageContent) {
+            message.delete();
+            message.channel.send(messageContent);
+        }
+
+        function err() {
+            message.reply(`Me desculpe, mas não posso enviar links!`);
+        }
+
     }
-
-    function err() {
-        message.reply(`Me desculpe, mas não posso enviar links!`);
-    }
-
 }
 
 module.exports.help = {
